pyproject.toml
[tool.poetry]
name = "stepik"
version = "0.1.0"
description = ""
authors = ["Mikhail Smetannikov <m.smetannikov@cryptology.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "1.11.17"
djangorestframework = "3.11"
pytest = "8.0.0"
pytest-django = "^4.11.1"
black = "24.8.0"
isort = "5.13.2"
celery = "5.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = [
    "core",
    "events",
    "finances",
    "orders",
]
skip_glob = ["infra/*", "venv/*", "*/migrations/*"]
line_length = 120
color_output = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
profile = "black"

[tool.pytest.ini_options]
pythonpath = "."
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", ]

[tool.black]
line-length = 120
force-exclude = '.*/migrations/.*'